requirements.txt
asgiref==3.8.1
Django==4.2.18
django-admin-rangefilter==0.13.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2==2.9.10
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
typing-extensions==4.12.2
